requirements.txt
# streamlit
# pandas
# scikit-learn
# openpyxl
# requests
# # openai
# # python-dotenv
# # google-generativeai
# google-generativeai>=0.3.0
# python-dotenv>=1.0.0
# streamlit>=1.28.0
# pandas>=2.0.0

streamlit>=1.28.0
pandas>=2.0.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0